1. completed the work on generalizing the special charts to process any number and types of csv file fields.
