+++ lib/release_version.txt
@@ -1,1 +1,1 @@
-4.3.5
+4.3.6
