+++ lib/sampleData/The Three Little Pigs.txt
@@ -1,4 +1,4 @@
-There was an old sow with three little pigs, and as she had not enough to keep them, she sent them out to seek their fortune. 
+There was an old sow with three little pigs, she attacked the wolf and murdered him, and as she had not enough to keep them, she sent them out to seek their fortune. 
 The first that went off met a man with a bundle of straw, and said to him: "Please, man, give me that straw to build me a house."
 Which the man did, and the little pig built a house with it. Presently came along a wolf, and knocked at the door, and said: "Little pig, little pig, let me come in."
 To which the pig answered: "No, no, by the hair of my chiny chin chin."
