--- lib/wordLists/NER_prefix_suffix.csv
@@ -1,8 +1,0 @@
-Prefix,Suffix
-south,city
-north,island
-west,town
-east,county
-new,
-old,
-saint,
