+++ src/NLP_menu_main.py
@@ -200,8 +200,8 @@ pydict["CoNLL table analyzer - Clause, noun, verb, function words frequencies"] 
 pydict["Statistics (csv files)"] = ["statistics_csv_main.py", 1]
 pydict["Statistics (txt files)"] = ["statistics_txt_ALL_main.py", 1] # ["style_analysis_main.py", 1]
 pydict["Co-Reference PRONOMINAL resolution (via Stanford CoreNLP)"] = ["parsers_annotators_main.py", 1]
-pydict["Co-Occurrences viewer"] = ["NGrams_CoOccurrences_main.py", 1]
-pydict["N-grams viewer"] = ["NGrams_CoOccurrences_main.py", 1]
+pydict["Co-Occurrences VIEWER"] = ["NGrams_CoOccurrences_main.py", 1]
+pydict["N-grams VIEWER"] = ["NGrams_CoOccurrences_main.py", 1]
 pydict["Data manipulation (ALL options GUI)"] = ["data_manipulation_main.py", 1]
 pydict["File checker (file content)"] = ["file_checker_converter_cleaner_main.py", 1]
 pydict["File checker (file content utf-8 encoding)"] = ["file_checker_converter_cleaner_main.py", 1]
@@ -214,7 +214,7 @@ pydict["File cleaner (Pronominal resolution via CoreNLP)"] = ["coreference_main.
 pydict["File classifier (dumb classifier via embedded date) (file name)"] = ["file_classifier_main.py", 1]
 pydict["File finder (file name)"] = ["file_manager_main.py", 1]
 pydict["File search (file content for words/collocations)"] = ["file_search_byWord_main.py", 1]
-pydict["File search (file content for n-grams & co-occurrences; N-grams viewer)"] = ["NGrams_CoOccurrences_main.py", 1]
+pydict["File search (file content for n-grams & co-occurrences)"] = ["NGrams_CoOccurrences_main.py", 1]
 pydict["File-type converter (csv, docx, pdf, rtf --> txt)"] = ["file_checker_converter_cleaner_main.py", 1]
 pydict["File matcher (file name)"] = ["file_matcher_main.py", 1]
 pydict["File merger (file content)"] = ["file_merger_main.py", 1]
@@ -225,6 +225,7 @@ pydict["Find non-related documents"] = ["social_science_research_main.py", 1]
 pydict["Excel charts"] = ["charts_Excel_main.py", 1]
 pydict["Animated time-dependent bar plot (Plotly)"] = ["data_visualization_1_main.py", 1]
 pydict["Boxplot"] = ["data_visualization_2_main.py", 1]
+pydict["Colormap"] = ["data_visualization_1_main.py", 1]
 pydict["Multiple bar charts"] = ["data_visualization_2_main.py", 1]
 pydict["Network graphs (Gephi)"] = ["data_visualization_1_main.py", 1]
 pydict["Sankey flowchart (Plotly)"] = ["data_visualization_1_main.py", 1]
@@ -242,7 +243,7 @@ pydict["Sample corpus (ALL options GUI)"] = ["sample_corpus_main.py", 1]
 pydict["File handler (ALL options GUI)"] = ["file_handler_ALL_main.py", 1]
 pydict["Search (ALL options GUI)"] = ["file_search_ALL_main.py", 1]
 pydict["Search CoNLL table"] = ["CoNLL_table_analyzer_main.py", 1]
-pydict["Search text file(s) for n-grams & co-occurrences (N-grams viewer)"] = ["NGrams_CoOccurrences_main.py", 1]
+pydict["Search text file(s) for n-grams & co-occurrences"] = ["NGrams_CoOccurrences_main.py", 1]
 pydict["Search text file(s) for words/collocations"] = ["file_search_byWord_main.py", 1]
 pydict["Sentence analysis (ALL options GUI)"] = ["sentence_analysis_main.py", 1]
 pydict["Sentence complexity"] = ["sentence_analysis_main.py", 1]
