+++ src/nominalization_main.py
@@ -12,405 +12,33 @@ import sys
 import GUI_util
 import IO_libraries_util
 
-if IO_libraries_util.install_all_Python_packages(GUI_util.window,"Nominalization",['tkinter','nltk','pywsd','wn','csv','re','os','collections'])==False:
-    sys.exit(0)
+# if IO_libraries_util.install_all_Python_packages(GUI_util.window,"Nominalization",['tkinter','nltk','pywsd','wn','csv','re','os','collections'])==False:
+#     sys.exit(0)
 
 import os
 import tkinter as tk
-import tkinter.messagebox as mb
-import nltk
-
-# check averaged_perceptron_tagger
-IO_libraries_util.import_nltk_resource(GUI_util.window,'taggers/averaged_perceptron_tagger','averaged_perceptron_tagger')
-# check punkt
-IO_libraries_util.import_nltk_resource(GUI_util.window,'tokenizers/punkt','punkt')
-# check WordNet
-IO_libraries_util.import_nltk_resource(GUI_util.window,'corpora/WordNet','WordNet')
-
-# from Stanza_functions_util import stanzaPipeLine, sent_tokenize_stanza
-# MUST use this version or code will break no longer true; pywsd~=1.2.4 pip install pywsd~=1.2.4; even try pip install pywsd=1.2.2
-#   or this version pip install pywsd==1.0.2
-# https://github.com/alvations/pywsd/issues/65
-# pywsd depends upon wn below; if the code breaks reinstall wn
-# pywsd Python word sense disambiguation
-#   https://pypi.org/project/pywsd/
-from pywsd import disambiguate
-from nltk.corpus import wordnet as wn
-# pip install wn==0.0.23
-import string
-import re
-from collections import Counter
-
-
 import IO_files_util
-import IO_csv_util
-import IO_user_interface_util
-import charts_util
 import GUI_IO_util
 
 # RUN section ______________________________________________________________________________________________________________________________________________________
 
-#first_section is to extract the word from "syns".
-# syns.name() of "intention" is "purpose.n.01". So "first_section" gets everything before the first period: "purpose"
-
-#params: sent > string
-#result #includes word='NO NOMINALIZATION'
-#count #includes word='NO NOMINALIZATION'
-#count1 #excludes word='NO NOMINALIZATION'
-def nominalized_verb_detection(docID,doc,sent,check_ending):
-    # sentences = tokenize.sent_tokenize(sent)
-    from Stanza_functions_util import stanzaPipeLine, sent_tokenize_stanza
-    sentences = sent_tokenize_stanza(stanzaPipeLine(sent))
-
-    result = []
-    result1 = []
-    verbs = []
-    true_word = []
-    false_word = []
-    # word count for the sentence
-    word_count = []
-    # number of nominalization in the sentence
-    nomi_count = []
-    sen_id = -1
-    # to print the sentence in output
-    sentence = []
-    # to print the nominalizations in each sentence
-    nomi_sen = []
-    nomi_sen_ = ""
-    def is_pos(s, pos):
-        # print(s)
-        return s.split('.')[1] == pos
-    for each_sen in sentences:
-        sen_id += 1
-        nomi_count.append(0)
-        word_count.append(0)
-        sentence.append(each_sen)
-        words_with_tags = disambiguate(each_sen)
-        for tup in words_with_tags:
-            word, syns = tup
-            if (word in string.punctuation) or (word == "\"") or (word[0] == "\'") or (word[0] == "`"):
-                continue
-            word_count[sen_id] += 1
-            derivationals = []
-            word = word.lower()
-            if word in true_word:
-                nomi_count[sen_id] += 1
-                if nomi_sen_ == "":
-                    nomi_sen_ = word
-                else:
-                    nomi_sen_ = nomi_sen_ + "; " + word
-                noun_cnt[word] += 1
-                nominalized_cnt[word] += 1
-                continue
-            if word in false_word:
-                noun_cnt[word] += 1
-                continue
-            if syns:
-                #look at only nouns
-                if not is_pos(syns.name(), 'n'):
-                    # TODO do not save; leads to huge file
-                    result.append([word, '', False])
-                    false_word.append(word)
-                    noun_cnt[word] += 1
-                    continue
-                if wn.lemmas(word):
-                    for lemma in wn.lemmas(word):
-                        derive = lemma.derivationally_related_forms()
-                        if derive not in derivationals and derive:
-                            derivationals.append(derive)
-                else:
-                    try:
-                        derivationals = syns.lemmas()[0].derivationally_related_forms()
-                    except:
-                        pass
-                stem = first_section.match(str(syns.name())).group(1)
-                found = False
-                for deriv in derivationals:
-                    if is_pos(str(deriv), 'v'):
-                        # original deriv_str = str(deriv)[7:-3].split('.')[3]
-                        # deriv is a list with typically one item
-                        #   [Lemma('construct.v.01.construct)
-                        # sometimes the list can have multiple items
-                        #   deriv [Lemma('dramatize.v.02.dramatize), Lemma('dramatize.v.02.dramatise')]
-                        #   when multiple items are given taken the first in the list
-                        #   deriv[0]
-                        try:
-                            deriv_str = str(deriv[0])[7:-2].split('.')[3]
-                        except:
-                            deriv_str = str([deriv][0])[7:-2].split('.')[3]
-                        if word=='lights':
-                            print('wrong')
-                        print('   NOUN:', word, ' VERB:',deriv_str)
-                        try:
-                            deriv_str = str(deriv[0])[7:-2].split('.')[3]
-                        except:
-                            continue
-                        # deriv_str is now the verb that is being lemmatized
-                        if check_ending:
-                            if not 'ent' in word[-3:] and not 'ing' in word[-3:] and not 'ion' in word[-3:] and \
-                                    not 'ance' in word[-4:] and not 'ence' in word[-4:]:
-                                continue
-                        result.append([word, deriv_str, True])
-                        verbs.append(deriv_str)
-                        true_word.append(word)
-                        noun_cnt[word] += 1
-                        if nomi_sen_ == "":
-                            nomi_sen_ = word
-                        else:
-                            nomi_sen_ = nomi_sen_ + "; " + word
-                        nominalized_cnt[word] += 1
-                        found = True
-                        break
-                if found:
-                    nomi_count[sen_id] += 1
-                    continue
-                else:
-                    # TODO do not save; leads to a huge file
-                    result.append([word, '', False]) #includes word='NO NOMINALIZATION'
-                    noun_cnt[word] += 1
-        nomi_sen.append(nomi_sen_)
-        nomi_sen_ = ""
-    for i in range(sen_id+1):
-        #['Document ID', 'Document', 'Sentence ID', 'Sentence', 'Number of words in sentence', 'Nominalized verbs','Number of nominalizations in sentence', 'Percentage of nominalizations in sentence'])
-        if word_count[i]>0:
-            # result1.append([docID, IO_csv_util.dressFilenameForCSVHyperlink(doc), i+1, sentence[i], word_count[i], nomi_sen[i], nomi_count[i],
-                              #                  100.0*nomi_count[i]/word_count[i]])
-
-            if nomi_count[i]>0:
-                result1.append([word_count[i], nomi_sen[i], nomi_count[i], 100.0 * nomi_count[i] / word_count[i],
-                           i + 1, sentence[i], docID, IO_csv_util.dressFilenameForCSVHyperlink(doc)])
-        else:
-            # result1.append([docID, IO_csv_util.dressFilenameForCSVHyperlink(doc), i+1, sentence[i], word_count[i], nomi_sen[i], nomi_count[i]])
-            if nomi_count[i]>0:
-                result1.append(
-                    [word_count[i], nomi_sen[i], nomi_count[i], sentence[i], docID, IO_csv_util.dressFilenameForCSVHyperlink(doc) ])
-    # print(result1)
-    # result contains a list of each word TRUE/FALSE values for nominalization
-    # result1 contains a list of docID, docName, sentence...
-    return result, result1
-
-def list_to_csv(outputFilename, lists):
-    # for list_ in lists:
-    #     word, bool = tup_
-    #     bool = str(bool)
-    #     outputFilename.write(word)
-    #     outputFilename.write(',')
-    #     outputFilename.write(bool)
-    #     outputFilename.write('\n')
-    IO_csv_util.list_to_csv(GUI_util.window,lists,outputFilename,colnum=0)
-
-def write_dir_csv(outputFilename, lists, file_name):
-    parts = file_name.split(os.path.sep)
-    file_name = parts[-1]
-    for list_ in lists:
-        list_.append(file_name)
-    IO_csv_util.list_to_csv(GUI_util.window,lists,outputFilename,colnum=0)
-
 def run(inputFilename,inputDir, outputDir,openOutputFiles,createCharts,chartPackage,doNotListIndividualFiles, check_ending):
-    global first_section, noun_cnt, nominalized_cnt
 
-    first_section = re.compile("^(.+?)\.")
-    noun_cnt = Counter()
-    nominalized_cnt = Counter()
-    filesToOpen = []  # Store all files that are to be opened once finished
+    filesToOpen = []
 
     if GUI_util.setup_IO_menu_var.get() == 'Default I/O configuration':
         config_filename = 'NLP_default_IO_config.csv'
     else:
         config_filename = scriptName.replace('main.py', 'config.csv')
 
-    # create a subdirectory of the output directory
-    outputDir = IO_files_util.make_output_subdirectory(inputFilename, inputDir, outputDir, label='NOM',
-                                                       silent=False)
-    if outputDir == '':
-        return
+    import nominalization_util
+    outputFiles = nominalization_util.nominalization(inputFilename,inputDir, outputDir, config_filename, openOutputFiles,createCharts,chartPackage,doNotListIndividualFiles, check_ending)
 
-    if __name__ == '__main__':
-        nltk.data.path.append('./nltk_data')
-
-        inputDocs = IO_files_util.getFileList(inputFilename, inputDir, fileType='.txt', silent=False,
-                                              configFileName=config_filename)
-        nDocs = len(inputDocs)
-        if nDocs == 0:
-            return filesToOpen
-
-        startTime=IO_user_interface_util.timed_alert(GUI_util.window,2000,'Analysis start', 'Started running Nominalization at',
-                                           True, '', True, '', False)
-
-        #add all into a sum
-        result_dir = []
-        docID=0
-        result2 = []
-        result_dir2=[]
-        counter_nominalized_list = []
-        counter_nominalized_list.append(['Nominalized verb', 'Frequency'])
-        counter_noun_list = []
-        counter_noun_list.append(['Noun', 'Frequency'])
-
-        for doc in inputDocs:
-
-            docID=docID+1
-            head, tail = os.path.split(doc)
-            print("Processing file " + str(docID) + "/" + str(nDocs) + ' ' + tail)
-            #open the doc and create the list of result (words, T/F)
-            fin = open(doc, 'r',encoding='utf-8',errors='ignore')
-            # result contains for each word the False/True nominalization boolean
-            # result1 contains the sentence and nominalized values for a specific document
-            result, result1 = nominalized_verb_detection(docID,doc,fin.read(),check_ending)
-            # result2 contains the sentence and nominalized values for all documents
-            result2.extend(result1)
-            fin.close()
-
-            # list all verbs as TRUE/FALSE if nominalized
-            for word, verb, boolean in result:
-                if boolean:
-                    # export only TRUE values
-                    if check_ending:
-                        if not 'ent' in word[-3:] and not 'ing' in word[-3:] and not 'ion' in word[-3:] and \
-                                not 'ance' in word[-4:] and not 'ence' in word[-4:]:
-                            continue
-                    result_dir.append([word, verb, boolean, docID, IO_csv_util.dressFilenameForCSVHyperlink(doc)])
-                else:
-                    continue
-            result_dir2.extend(result_dir)
-
-            if len(inputDir) > 0:
-                fname = os.path.basename(os.path.normpath(inputDir))+"_dir"
-            else:
-                fname=doc
-
-            # used for both individual files and directories
-            # outputFilename_bySentenceIndex = IO_files_util.generate_output_file_name(fname, '', outputDir,
-            #                                                                           '.csv', 'NOM', 'sent', '', '',
-            #                                                                           '', False, True)
-            #
-            outputFilename_bySentenceIndex = IO_files_util.generate_output_file_name(fname, '', outputDir,
-                                                                                      '.csv', 'NOM', 'sent')
-            if len(inputDir) == 0 or doNotListIndividualFiles == False:
-                counter_nominalized_list = []
-                counter_noun_list = []
-                # refresh the headers
-                counter_nominalized_list.insert(0,['Nominalized verb', 'Frequency'])
-                counter_noun_list.insert(0,['Noun', 'Frequency'])
-
-                result1.insert(0, ['Number of words in sentence', 'Nominalized verbs',
-                                   'Number of nominalizations in sentence',
-                                   'Percentage of nominalizations in sentence',
-                                   'Sentence ID', 'Sentence', 'Document ID', 'Document'])
-
-                # compute frequency of most common nominalized verbs
-                for word, freq in nominalized_cnt.most_common():
-                    counter_nominalized_list.append([word, freq])
-
-                # compute frequency of most common nouns
-                for word, freq in noun_cnt.most_common():
-                    counter_noun_list.append([word, freq])
-
-                head, fname=os.path.split(doc)
-                fname=fname[:-4]
-
-                outputFilename_noun_frequencies = IO_files_util.generate_output_file_name(fname, inputDir, outputDir, '.csv', 'NOM',
-                                                                                'noun_freq')
-                filesToOpen.append(outputFilename_noun_frequencies)
-                outputFilename_nominalized_frequencies = IO_files_util.generate_output_file_name(fname,
-                                                                                inputDir, outputDir, '.csv', 'NOM',
-                                                                                 'nominal_freq')
-
-
-                filesToOpen.append(outputFilename_nominalized_frequencies)
-
-                # export nominalized verbs
-                list_to_csv(outputFilename_nominalized_frequencies, counter_nominalized_list)
-
-                # export nouns
-                list_to_csv(outputFilename_noun_frequencies, counter_noun_list)
-
-                outputFilename_TRUE_FALSE = IO_files_util.generate_output_file_name(fname + '_TRUE_FALSE', '', outputDir,
-                                                                               '.csv', 'NOM')
-
-                # TODO this leads to a huge file when processing a directory; comment for now
-                filesToOpen.append(outputFilename_TRUE_FALSE)
-                # this will export both True and False
-                # list_to_csv(outputFilename_TRUE_FALSE, result)
-                result_dir2.insert(0, ["Noun", "Verb", "Is nominalized", "Document ID", "Document"])
-                list_to_csv(outputFilename_TRUE_FALSE, result_dir2)
-
-                list_to_csv(outputFilename_bySentenceIndex, result1)
-                filesToOpen.append(outputFilename_bySentenceIndex)
-
-        if len(inputDir)>0 and doNotListIndividualFiles == True:
-            # outputFilename_TRUE_FALSE_dir = IO_files_util.generate_output_file_name(fname + '_TRUE_FALSE', '', outputDir, '.csv', 'NOM', '', '', '', '', False, True)
-            outputFilename_TRUE_FALSE_dir = IO_files_util.generate_output_file_name(fname + '_TRUE_FALSE', '', outputDir, '.csv', 'NOM')
-            outputFilename_dir_noun_frequencies=IO_files_util.generate_output_file_name(fname, '', outputDir, '.csv', 'NOM', 'noun_freq', '', '', '', False, True)
-            # outputFilename_dir_nominalized_frequencies=IO_files_util.generate_output_file_name(fname, '', outputDir, '.csv', 'NOM', 'nominal_freq', '', '', '', False, True)
-            outputFilename_dir_nominalized_frequencies=IO_files_util.generate_output_file_name(fname, '', outputDir, '.csv', 'NOM', 'nominal_freq')
-
-            result2.insert(0, ['Number of words in sentence', 'Nominalized verbs',
-                               'Number of nominalizations in sentence', 'Percentage of nominalizations in sentence','Sentence ID', 'Sentence', 'Document ID', 'Document'])
-            list_to_csv(outputFilename_bySentenceIndex, result2)
-            filesToOpen.append(outputFilename_bySentenceIndex)
-
-            # list all verbs as TRUE/FALSE if nominalized
-            # TODO  this leads to a huge file when processing a directory; comment for now
-            result_dir2.insert(0, ["Noun", "Verb", "Is nominalized", "Document ID", "Document"])
-            list_to_csv(outputFilename_TRUE_FALSE_dir, result_dir2)
-            filesToOpen.append(outputFilename_TRUE_FALSE_dir)
-
-            counter_noun_list = []
-            counter_noun_list.append(['Noun','Frequency'])
-            for word, freq in noun_cnt.most_common():
-                counter_noun_list.append([word, freq])
-            list_to_csv(outputFilename_dir_noun_frequencies, counter_noun_list)
-            filesToOpen.append(outputFilename_dir_noun_frequencies)
-
-            counter_nominalized_list = []
-            counter_nominalized_list.append(['Nominalized verb','Frequency'])
-            for word, freq in nominalized_cnt.most_common():
-                counter_nominalized_list.append([word, freq])
-            list_to_csv(outputFilename_dir_nominalized_frequencies, counter_nominalized_list)
-            filesToOpen.append(outputFilename_dir_nominalized_frequencies)
-
-            if createCharts == True:
-                # bar chart of nominalized verbs
-
-                inputFilename = outputFilename_dir_nominalized_frequencies
-                columns_to_be_plotted_xAxis=[]
-                columns_to_be_plotted_yAxis=[[0, 1]]
-
-                outputFiles = charts_util.run_all(columns_to_be_plotted_yAxis, inputFilename, outputDir,
-                                                                 outputFileLabel='NOM_verb',
-                                                                 chartPackage=chartPackage,
-                                                                 chart_type_list=['bar'],
-                                                                 chart_title='Nominalized Verbs',
-                                                                 column_xAxis_label_var='column_xAxis_label',
-                                                                 hover_info_column_list=[],
-                                                                 count_var=0)
-                if outputFiles!=None:
-                    if isinstance(outputFiles, str):
-                        filesToOpen.append(outputFiles)
-                    else:
-                        filesToOpen.extend(outputFiles)
-
-                inputFilename = outputFilename_dir_noun_frequencies
-                columns_to_be_plotted_xAxis=[]
-                columns_to_be_plotted_yAxis=[[0, 1]]
-
-                outputFiles = charts_util.run_all(columns_to_be_plotted_yAxis, inputFilename, outputDir,
-                                                                 outputFileLabel='NOM_noun',
-                                                                 chartPackage=chartPackage,
-                                                                 chart_type_list=['bar'],
-                                                                 chart_title='Nouns',
-                                                                 column_xAxis_label_var='column_xAxis_label',
-                                                                 hover_info_column_list=[],
-                                                                 count_var=0)
-
-                if outputFiles!=None:
-                    if isinstance(outputFiles, str):
-                        filesToOpen.append(outputFiles)
-                    else:
-                        filesToOpen.extend(outputFiles)
-
-    IO_user_interface_util.timed_alert(GUI_util.window,2000,'Analysis end', 'Finished running Nominalization at', True, '', True, startTime)
+    if outputFiles!=None:
+        if isinstance(outputFiles, str):
+            filesToOpen.append(outputFiles)
+        else:
+            filesToOpen.extend(outputFiles)
 
     if openOutputFiles == 1:
         IO_files_util.OpenOutputFiles(GUI_util.window, openOutputFiles, filesToOpen, outputDir, scriptName)
@@ -427,6 +55,7 @@ run_script_command=lambda: run(GUI_util.inputFilename.get(),
                                 check_nom_verb_ending_var.get())
 
 GUI_util.run_button.configure(command=run_script_command)
+
 
 # GUI section ______________________________________________________________________________________________________________________________________________________
